chore(deps): update schema digest to b812391

+++ src/models/payment_request_create_params_payment_request_lines_inner.rs
@@ -21,7 +21,7 @@ pub struct PaymentRequestCreateParamsPaymentRequestLinesInner {
     pub description: Option<String>,
     /// 金額
     #[serde(rename = "amount")]
-    pub amount: i32,
+    pub amount: i64,
     /// 勘定科目ID
     #[serde(rename = "account_item_id", skip_serializing_if = "Option::is_none")]
     pub account_item_id: Option<i32>,
@@ -49,7 +49,7 @@ pub struct PaymentRequestCreateParamsPaymentRequestLinesInner {
 }
 
 impl PaymentRequestCreateParamsPaymentRequestLinesInner {
-    pub fn new(amount: i32) -> PaymentRequestCreateParamsPaymentRequestLinesInner {
+    pub fn new(amount: i64) -> PaymentRequestCreateParamsPaymentRequestLinesInner {
         PaymentRequestCreateParamsPaymentRequestLinesInner {
             line_type: None,
             description: None,
